# Calcpad.WebApi/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Calcpad.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/calcpad-file/code/{uniqueId}/exists": {
      "get": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "PresignedObject",
        "description": "verify if a file with the given sha256 exists\r\nif exists, return the file id, otherwise return empty string\r\nif the file exists, also verify if the filePath exists for the user, if not, add a new accessor",
        "parameters": [
          {
            "name": "uniqueId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calcpad-file": {
      "post": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "UploadCalcpadCodeFile",
        "description": "upload a calcpad file and return the file id\r\nif the file content is changed, you must upload it again with same fileName",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "UniqueId": {
                    "type": "string",
                    "description": "unique id for the file, can be generated by client\r\nthis id should be stored in database to identify the file in client or other service"
                  },
                  "FileName": {
                    "type": "string",
                    "description": "this file name is only for display, the real saved file name is generated by server\r\nprovided by client"
                  },
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "IsPublic": {
                    "type": "boolean",
                    "description": "true when the file is an image file, false when the file is a calcpad code file"
                  }
                }
              },
              "encoding": {
                "UniqueId": {
                  "style": "form"
                },
                "FileName": {
                  "style": "form"
                },
                "File": {
                  "style": "form"
                },
                "IsPublic": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calcpad-file/copy": {
      "post": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "CopyCalcpadCodeFile",
        "description": "copy cpd file from fromId to toId",
        "parameters": [
          {
            "name": "fromId",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toId",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calcpad-file/uniqueId": {
      "delete": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "DeleteCalcpadFile",
        "description": "Delete calcpad file",
        "parameters": [
          {
            "name": "uniqueId",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calcpad-file/stream/public/{fileId}": {
      "get": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "GetFileStream",
        "description": "获取文件流",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calcpad-file/input-form": {
      "get": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "CompileToInputForm",
        "description": "compile to input form",
        "parameters": [
          {
            "name": "uniqueId",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "simplify",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calcpad-file/{uniqueId}/calculate-html": {
      "post": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "RunCalculations",
        "description": "run calculations and return the result as html",
        "parameters": [
          {
            "name": "uniqueId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RunCalculationData"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCalculationData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCalculationData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RunCalculationData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calcpad-file/{uniqueId}/origin": {
      "get": {
        "tags": [
          "CalcpadFile"
        ],
        "summary": "DownloadCpdFile",
        "description": "download cpd file and includes",
        "parameters": [
          {
            "name": "uniqueId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/user/sign-in": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "SignIn",
        "description": "sign in to get the token",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SignInData"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SignInData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RunCalculationData": {
        "type": "object",
        "properties": {
          "cpdUid": {
            "type": "string",
            "nullable": true
          },
          "inputFields": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "输入字段",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SignInData": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}